# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API Sistema de Farmacia",
        "description": "Sistema para la gestión de medicamentos y movimientos de stock",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "Usuarios",
            "description": "Gestión de usuarios en el sistema"
        },
        {
            "name": "Medicamentos",
            "description": "Gestión de medicamentos en el sistema"
        },
        {
            "name": "Movimientos",
            "description": "Gestión de movimientos de stock en el sistema"
        }
    ],
    "paths": {
        "/usuarios/{id}/eliminar": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Realiza una eliminación lógica de un usuario",
                "operationId": "eliminarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/medicamentos/{id}": {
            "get": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Buscar un medicamento por su ID",
                "operationId": "buscar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medicamento"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Actualizar el stock de un medicamento",
                "operationId": "actualizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Medicamento"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medicamento"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Eliminar un medicamento",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar todos los usuarios",
                "operationId": "listarUsuarios",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "crearUsuario",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movimientos/salida": {
            "post": {
                "tags": [
                    "Movimientos"
                ],
                "summary": "Registrar una salida de stock",
                "operationId": "registrarSalida",
                "parameters": [
                    {
                        "name": "medicamentoId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "cantidad",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movimiento"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movimientos/entrada": {
            "post": {
                "tags": [
                    "Movimientos"
                ],
                "summary": "Registrar una entrada de stock",
                "operationId": "registrarEntrada",
                "parameters": [
                    {
                        "name": "medicamentoId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "cantidad",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movimiento"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/medicamentos": {
            "get": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Listar todos los medicamentos",
                "operationId": "listar",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Medicamento"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Crear un nuevo medicamento",
                "operationId": "crear",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Medicamento"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medicamento"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/medicamentos/{id}/stock": {
            "patch": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Actualizar el stock de un medicamento",
                "operationId": "actualizarStock",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medicamento"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Buscar un usuario por su ID",
                "operationId": "buscarPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movimientos/{id}": {
            "delete": {
                "tags": [
                    "Movimientos"
                ],
                "summary": "Eliminar un movimiento",
                "operationId": "eliminarMovimiento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Medicamento": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "COMPRIMIDOS",
                            "JARABE"
                        ]
                    },
                    "marca": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "Usuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "rol": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "OPERADOR"
                        ]
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "bloqueado": {
                        "type": "boolean"
                    }
                }
            },
            "Movimiento": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "ENTRADA",
                            "SALIDA"
                        ]
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "medicamento": {
                        "$ref": "#/components/schemas/Medicamento"
                    }
                }
            }
        }
    }
}